feat: created a private route component and created a create case form

+++ src/services/caseCall.jsx
@@ -8,6 +8,7 @@ export const getAllCases = async (token) => {
                 Authorization: `Bearer ${token}`,
             },
         };
+        if(!token) {return }
         const response = await axios.get("http://localhost:3000/api/cases", config);
         return response;
     } catch (error) {
@@ -24,7 +25,7 @@ export const getCaseById = async (id,token) => {
                 Authorization: `Bearer ${token}`,
             },
         };
-
+        if(!id || !token) {return }
         const response = await axios.get(`http://localhost:3000/api/cases/${id}`, config);
         return response;
     } catch (error) {
@@ -40,6 +41,7 @@ export const editCaseCall = async (id, caseData, token) => {
                 Authorization: `Bearer ${token}`,
             },
         };
+        if(!id || !token ||!caseData) {return }
         const response = await axios.put(`http://localhost:3000/api/cases/${id}`, caseData, config);
         return response;
     } catch (error) {
@@ -55,6 +57,7 @@ export const deleteCaseCall = async (id, token) => {
                 Authorization: `Bearer ${token}`,
             },
         };
+        if(!id || !token) {return }
         const response = await axios.delete(`http://localhost:3000/api/cases/${id}`, config);
         return response;
     } catch (error) {
@@ -70,7 +73,25 @@ export const createCaseCall = async (caseData, token) => {
                 Authorization: `Bearer ${token}`,
             },
         };
+        if(!caseData || !token) {return }
         const response = await axios.post("http://localhost:3000/api/cases", caseData, config);
+        return response;
+    } catch (error) {
+        return error.response;
+    }
+}
+
+//Get Case By tech ID
+export const getCaseByUserId = async (token) => {
+    try {
+        const config = {
+            headers: {
+                Authorization: `Bearer ${token}`,
+            },
+        };
+        if(!token) {return }
+        const response = await axios.get(`http://localhost:3000/api/cases/technician/`, config);
+        
         return response;
     } catch (error) {
         return error.response;
